Check if IP are in use and only delete them then

+++ scripts/gcp-assign-vip.sh
@@ -15,11 +15,14 @@ while $internal_status || $external_status; do
     INTERNAL_INSTANCE_NAME=$(gcloud compute addresses describe ${internal_vip} --region=${INTERNAL_INSTANCE_REGION} --format='get(users[0])'|awk -F'/' '{print $NF}')
     INTERNAL_INSTANCE_ZONE=$(gcloud compute instances list --filter="name=${INTERNAL_INSTANCE_NAME}"|grep ${INTERNAL_INSTANCE_NAME}|awk '{print $2}')
     INTERNAL_INSTANCE_STATUS=$(gcloud compute instances describe --zone=${INTERNAL_INSTANCE_ZONE} $INTERNAL_INSTANCE_NAME --format='get(status)')
-    #Update the alias from the terminated instance to null
-    until gcloud compute instances network-interfaces update $INTERNAL_INSTANCE_NAME --zone $INTERNAL_INSTANCE_ZONE --aliases "" &>> /var/log/gcp-failoverd/default.log; do
-      echo "Trying to update the alias from $INTERNAL_INSTANCE_NAME to null"
-      sleep 2
-    done
+    if [[ $INTERNAL_IP_STATUS == "IN_USE" ]];
+    then
+      #Update the alias from the terminated instance to null
+      until gcloud compute instances network-interfaces update $INTERNAL_INSTANCE_NAME --zone $INTERNAL_INSTANCE_ZONE --aliases "" &>> /var/log/gcp-failoverd/default.log; do
+        echo "Trying to update the alias from $INTERNAL_INSTANCE_NAME to null"
+        sleep 2
+      done
+    fi
     # Assign IP aliases to me because now I am the MASTER!
     until gcloud compute instances network-interfaces update $(hostname) --zone $ZONE --aliases "${INTERNAL_IP}/32" &>> /var/log/gcp-failoverd/default.log; do
       echo "Trying to assign IP aliases to me because now I am the MASTER!"
@@ -39,11 +42,14 @@ while $internal_status || $external_status; do
     EXTERNAL_INSTANCE_ZONE=$(gcloud compute instances list --filter="name=${EXTERNAL_INSTANCE_NAME}"|grep ${EXTERNAL_INSTANCE_NAME}|awk '{print $2}')
     EXTERNAL_INSTANCE_STATUS=$(gcloud compute instances describe --zone=${EXTERNAL_INSTANCE_ZONE} $EXTERNAL_INSTANCE_NAME --format='get(status)')
     EXTERNAL_ACCESS_CONFIG=$(gcloud compute instances describe --zone=${EXTERNAL_INSTANCE_ZONE} $EXTERNAL_INSTANCE_NAME --format='get(networkInterfaces[0].accessConfigs[0].name)')
-    #Delete the access config from the terminated node
-    until gcloud compute instances delete-access-config --zone=${EXTERNAL_INSTANCE_ZONE} $EXTERNAL_INSTANCE_NAME --access-config-name=${EXTERNAL_ACCESS_CONFIG} &>> /var/log/gcp-failoverd/default.log; do
-      echo "Trying to Delete the access config from $INTERNAL_INSTANCE_NAME"
-      sleep 2
-    done
+    if [[ $EXTERNAL_IP_STATUS == "IN_USE" ]];
+    then
+      #Delete the access config from the terminated node
+      until gcloud compute instances delete-access-config --zone=${EXTERNAL_INSTANCE_ZONE} $EXTERNAL_INSTANCE_NAME --access-config-name=${EXTERNAL_ACCESS_CONFIG} &>> /var/log/gcp-failoverd/default.log; do
+        echo "Trying to Delete the access config from $INTERNAL_INSTANCE_NAME"
+        sleep 2
+      done
+    fi
     # Assign Access Config to me because now I am the MASTER!
     until gcloud compute instances add-access-config $(hostname) --zone $ZONE --access-config-name "$(hostname)-access-config" --address $EXTERNAL_IP &>> /var/log/gcp-failoverd/default.log; do
       echo "Trying to assign IP access config to me because now I am the MASTER!"
